Resolved issue with cursor improper placement
Cursor wasnt showing in the correct place when starting brackets and the
cursor was set in a line starting with tabs.  The issue was caused due
to applying the css rules after the document was already loaded, causing
codemirror to do tab calculation based on dimensions with padding that
was about to change.  The issue was resolved by adding a default padding
to document lines of 2px rather than 15px as soon as themes is loaded. I
am overriding 15px because that's too much space from the beginning of
each line