Adjustments in styles in resest and visual-studio theme
Adjusted reset.css to have the value that a future version of CodeMirror
will have.  Also adjusted the size of the scrollbar to be a tad bigger
for easier grabbing/dragging.  In Visual Studio theme I changed active
lines from border to outlines so that active lines don't look so
crowded.  Also, to follow what the Eclipse theme is doing