Addressed highlighting concerns in every theme.
The update is intended to be verbose so as to provide as much opportunity for individual tweaking as possible. 

In every case opacity has been added to the background-color to dim (or enhance) the brightness of the highlight. 

The pixie theme currently runs into: background-color: #222222 !important; in the pixie-css:135 meaning it doesnt currently display.